Ref: ParticipantCount(enum) 파일로 분리
