+++ PokerGameApp/PokerGameApp/ParticipantCount.swift
@@ -1,0 +1,15 @@
+//
+//  ParticipantCount.swift
+//  PokerGameApp
+//
+//  Created by jsj on 2022/02/25.
+//
+
+import Foundation
+
+enum ParticipantCount: Int {
+    case one = 1
+    case two = 2
+    case three = 3
+    case four = 4
+}
